WIP: running parallell unit tests and integrations
