+++ .gitignore
@@ -10,6 +10,7 @@ dist/*.gz
 # Other ignores
 tmp/
 .idea/
+.eslintcache
 
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
