--- addons/Dexie.Observable/test/karma-env.js
@@ -1,3 +1,0 @@
-QUnit.config.autostart = false;
-window.workerImports = ['../dist/dexie.js'];
-window.workerSource = '/base/test/worker.js';
