--- addons/Dexie.Observable/test/test-observable-dexie-tests.html
@@ -1,22 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Dexie Unit tests with Dexie.Observable applied</title>
-  <link rel="stylesheet" href="../../../node_modules/qunitjs/qunit/qunit.css">
-</head>
-<body>
-    <div id="qunit"></div>
-    <div id="qunit-fixture"></div>
-    <script src="../../../test/babel-polyfill/polyfill.min.js"></script>
-    <script>
-        // Hack for making the WebWorker test behave also here...
-        window.workerSource = "../../../test/worker.js";
-        window.workerImports = ["../dist/dexie.js", "../addons/Dexie.Observable/dist/dexie-observable.js"];
-    </script>
-    <script src="../../../node_modules/qunitjs/qunit/qunit.js"></script>
-    <script src="../../../dist/dexie.js"></script>
-    <script src="../dist/dexie-observable.js"></script>
-    <script src="../../../test/bundle.js"></script>
-</body>
-</html>
