--- addons/Dexie.Observable/test/test-observable.html
@@ -1,16 +1,0 @@
-﻿<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title>Dexie.Observable Unit tests</title>
-    <link rel="stylesheet" href="../../../node_modules/qunitjs/qunit/qunit.css">
-</head>
-<body>
-    <div id="qunit"></div>
-    <div id="qunit-fixture"></div>
-    <script src="../../../node_modules/qunitjs/qunit/qunit.js"></script>
-    <script src="../../../dist/dexie.js"></script>
-    <script src="../dist/dexie-observable.js"></script>
-    <script src="tests-observable-misc.js"></script>
-</body>
-</html>
