+++ addons/Dexie.Observable/test/typings/test-typings.ts
