+++ addons/Dexie.Observable/test/typings/tsconfig.json
