--- addons/Dexie.Observable/test/unit/karma-env.js
@@ -1,1 +1,0 @@
-QUnit.config.autostart = false;
