+++ addons/Dexie.Observable/test/unit/karma.conf.js
@@ -1,48 +1,16 @@
-module.exports = function(config) {
-  const configuration = {
+// Include common configuration
+const karmaCommon = require('../../../../test/karma.common');
+
+module.exports = function (config) {
+  config.set(Object.assign({}, karmaCommon, {
+    // Base path should point at the root 
     basePath: '../../../../',
-
-    frameworks: [
-      'qunit'
-    ],
-
-    reporters: [
-      'mocha'
-    ],
-
-    client: {
-      captureConsole: true
-    },
-
-    files: [
-      './test/babel-polyfill/polyfill.min.js',
-      'node_modules/qunitjs/qunit/qunit.js',
-      'test/karma-env.js',
+    files: karmaCommon.files.concat([
       'dist/dexie.js',
+      'addons/Dexie.Observable/dist/dexie-observable.js',
       'addons/Dexie.Observable/test/unit/bundle.js',
-      { pattern: '**/*.map', watched: false, included: false, served: true}
-    ],
-
-    port: 19144,
-    //captureTimeout: 30 * 1000,
-    //browserNoActivityTimeout: 10 * 60 * 1000,
-    colors: true,
-
-    // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
-    //logLevel: config.LOG_DEBUG,
-
-    browsers: [
-        'Chrome'
-    ],
-
-    plugins: [
-      'karma-qunit',
-      'karma-mocha-reporter',
-      'karma-chrome-launcher',
-      //'karma-firefox-launcher'
-    ]
-  };
-
-  config.set(configuration);
-};
+      { pattern: 'addons/Dexie.Observable/test/unit/*.map', watched: false, included: false },
+      { pattern: 'addons/Dexie.Observable/dist/*.map', watched: false, included: false }
+    ])
+  }));
+}
