+++ addons/Dexie.Observable/test/unit/run-unit-tests.html
@@ -11,6 +11,7 @@
     <script src="../../../../test/babel-polyfill/polyfill.min.js"></script>
     <script src="../../../../node_modules/qunitjs/qunit/qunit.js"></script>
     <script src="../../../../dist/dexie.js"></script>
+    <script src="../../dist/dexie-observable.js"></script>
     <script src="bundle.js"></script>
 </body>
 </html>
