+++ addons/Dexie.Observable/test/unit/travis.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -e
+echo "Installing dependencies for dexie-observable"
+npm install >/dev/null
+npm run build
+npm run test:unit
