+++ addons/Dexie.Observable/test/unit/unit-tests-all.js
@@ -3,3 +3,4 @@ import './hooks/tests-deleting.js';
 import './hooks/tests-updating.js';
 import './tests-override-open.js';
 import './tests-override-parse-stores-spec.js';
+import './tests-observable-misc';
