+++ test/karma.browserstack.js
@@ -1,18 +1,10 @@
-module.exports = function(config) {
-  var cfg = {};
-  require('./karma.conf')({set: function (x){cfg = x;}});
-  cfg.client = {
-    captureConsole: true
-  };
-  cfg.port = 19145;
-  cfg.browserStack = {
-      username: process.env.BROWSER_STACK_USERNAME,
-      accessKey: process.env.BROWSER_STACK_ACCESS_KEY
-  };
-  if (!cfg.browserStack.username)
-    throw new Error("You must provider username/key in the env variables BROWSER_STACK_USERNAME and BROWSER_STACK_ACCESS_KEY");
+module.exports = {
+  browserStack: {
+    username: process.env.BROWSER_STACK_USERNAME,
+    accessKey: process.env.BROWSER_STACK_ACCESS_KEY
+  },
 
-  cfg.customLaunchers = {
+  customLaunchers: {
     bs_firefox_latest: {
       base: 'BrowserStack',
       browser: 'firefox',
@@ -55,22 +47,5 @@ module.exports = function(config) {
       os: 'Windows',
       os_version: 10
     }
-  };
-
-  cfg.browsers = [
-    'bs_chrome',
-    'bs_chrome_latest',
-    'bs_firefox',
-    'bs_firefox_latest',
-    'bs_edge',
-    'bs_ie11'
-  ];
-
-  cfg.plugins = [
-      'karma-qunit',
-      'karma-mocha-reporter',
-      'karma-browserstack-launcher'
-  ];
-
-  config.set(cfg);
-};
+  }
+}
