+++ test/karma.conf.js
@@ -1,49 +1,14 @@
-module.exports = function(config) {
-  const configuration = {
+// Include common configuration
+const karmaCommon = require('./karma.common');
+
+module.exports = function (config) {
+  config.set(Object.assign({}, karmaCommon, {
+    // Base path should point at the root 
     basePath: '..',
-
-    frameworks: [
-      'qunit'
-    ],
-
-    reporters: [
-      'mocha'
-    ],
-
-    client: {
-      captureConsole: false
-    },
-
-    files: [
-      'test/babel-polyfill/polyfill.min.js',
-      'node_modules/qunitjs/qunit/qunit.js',
-      'test/karma-env.js',
+    files: karmaCommon.files.concat([
       'dist/dexie.js',
       'test/bundle.js',
-      { pattern: 'test/worker.js', included: false },
-      { pattern: '**/*.map', watched: false, included: false, served: true}
-    ],
-
-    port: 19144,
-    //captureTimeout: 30 * 1000,
-    browserNoActivityTimeout: 2 * 60 * 1000,
-    colors: true,
-
-    // level of logging
-    // possible values: config.LOG_DISABLE || config.LOG_ERROR || config.LOG_WARN || config.LOG_INFO || config.LOG_DEBUG
-    //logLevel: config.LOG_DEBUG,
-
-    browsers: [
-        'Chrome'
-    ],
-
-    plugins: [
-      'karma-qunit',
-      'karma-mocha-reporter',
-      'karma-chrome-launcher',
-      //'karma-firefox-launcher'
-    ]
-  };
-
-  config.set(configuration);
-};
+      { watched: true, included: false, served: true, pattern: 'test/worker.js' },
+    ])
+  }));
+}
