+++ test/travis.sh
@@ -1,0 +1,3 @@
+#!/bin/bash -e
+npm run test:typings
+npm run test:unit
